requirements.txt
ultralytics==8.3.0
torch
torchvision
roboflow
dvc
pandas
wandb
onnx>=1.14.0
onnxruntime
fastapi
uvicorn[standard]
pydantic
pytest
httpx
python-dotenv